requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
python-socketio==5.11.4
aiohttp==3.10.5
